+++ stroom-processor/stroom-processor-impl/src/main/java/stroom/processor/impl/ProcessorTaskManagerImpl.java
@@ -419,10 +419,12 @@ class ProcessorTaskManagerImpl implements ProcessorTaskManager {
         LOGGER.debug("doCreateTasks() - Starting");
 
         // Get an up to date list of all enabled stream processor filters.
-        LOGGER.trace("Getting enabled stream processor filters");
+        LOGGER.trace("Getting enabled non deleted processor filters");
         final ExpressionOperator expression = new Builder()
                 .addTerm(ProcessorFilterFields.PROCESSOR_ENABLED, Condition.EQUALS, true)
+                .addTerm(ProcessorFilterFields.PROCESSOR_DELETED, Condition.EQUALS, false)
                 .addTerm(ProcessorFilterFields.PROCESSOR_FILTER_ENABLED, Condition.EQUALS, true)
+                .addTerm(ProcessorFilterFields.PROCESSOR_FILTER_DELETED, Condition.EQUALS, false)
                 .build();
 
         final ExpressionCriteria findProcessorFilterCriteria = new ExpressionCriteria(expression);